Isolate object-bound/function implementation and related Karma tests.
Karma has no way to run tests for groups of files in complete isolation.
Testing different implementations for the same global function requires
sequential tests, so that we load first implementation and test it, then
load the second implementation and test it, and so on.

To achieve this we will create a separate Karma config for each
implementation and related tests.